@@ -6,28 +6,28 @@
 	"license": "MIT",
 	"private": true,
 	"devDependencies": {
-		"@babel/core": "^7.17.5",
+		"@babel/core": "^7.17.9",
 		"@babel/plugin-proposal-class-properties": "^7.16.7",
 		"@babel/preset-env": "^7.16.11",
 		"@babel/preset-react": "^7.16.7",
-		"babel-loader": "^8.2.3",
+		"babel-loader": "^8.2.5",
 		"clean-webpack-plugin": "^4.0.0",
-		"css-loader": "^6.6.0",
+		"css-loader": "^6.7.1",
 		"electron": "^17.1.0",
-		"electron-packager": "^15.4.0",
+		"electron-packager": "^15.5.1",
 		"file-loader": "^6.2.0",
 		"html-webpack-plugin": "^5.5.0",
 		"javascript-obfuscator": "^4.0.0",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
+		"react": "^18.0.0",
+		"react-dom": "^18.0.0",
 		"resolve-url-loader": "^5.0.0",
-		"sass": "^1.49.9",
+		"sass": "^1.50.1",
 		"sass-loader": "^12.6.0",
 		"style-loader": "^3.3.1",
 		"url-loader": "^4.1.1",
 		"webpack": "^5.72.0",
 		"webpack-cli": "^4.9.2",
-		"webpack-dev-server": "^4.7.4"
+		"webpack-dev-server": "^4.8.1"
 	},
 	"scripts": {
 		"build": "webpack --mode production",
@@ -38,12 +38,12 @@
 	},
 	"dependencies": {
 		"@ant-design/icons": "^4.7.0",
-		"antd": "^4.18.8",
+		"antd": "^4.19.5",
 		"antd-img-crop": "^4.2.0",
 		"axios": "^0.26.1",
 		"classnames": "^2.3.1",
 		"electron-store": "^8.0.1",
 		"jwt-decode": "^3.1.2",
-		"redux": "^4.1.2"
+		"redux": "^4.2.0"
 	}
 }
