@@ -1,5 +1,6 @@
 import React from "react";
-import ReactDOM from 'react-dom';
-import App from "./App";
+import { createRoot } from 'react-dom/client'
+import App from './App';
 
-ReactDOM.render(<App />, document.getElementById('root'))
+const root = createRoot(document.getElementById('root'))
+root.render(<App />)
